Storybook stories moved to component folders
