+++ config/storybook/config.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { configure } from "@storybook/vue";
 
-const req = require.context("../../src/stories", true, /.stories.js$/);
+const req = require.context("../../src/", true, /.stories.js$/);
 
 function loadStories() {
   req.keys().forEach(filename => req(filename));
