+++ src/components/molecules/SfCallToAction/SfCallToAction.js
@@ -2,6 +2,9 @@ import SfButton from "../../atoms/SfButton/SfButton.vue";
 
 export default {
   name: "SfCallToAction",
+  props: {
+    buttonText: String
+  },
   components: {
     SfButton
   }
